% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.connection.R
\name{get.connection}
\alias{get.connection}
\title{Get Connection Class for Documenter}
\usage{
get.connection(server, catalog)
}
\arguments{
\item{server}{Server (Instance) Name}

\item{catalog}{Database Name}
}
\description{
This function returns an ODBC connection class to a T-SQL Server.
}
\examples{
cnn <- get.connection('<server_name>','<database_name>')
}
\keyword{tsql}
