% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.documentation.R
\name{get.documentation}
\alias{get.documentation}
\title{Get Documentation from T-SQL Metadata}
\usage{
get.documentation(connection, output.directory, output.file)
}
\arguments{
\item{connection}{ODBC connection class}

\item{output.directory}{The directory where the documentation will land.}

\item{output.file}{The output name of the documentation file.}
}
\description{
This function used R markdown to knit a PDF document of the data definition language (DDL) of any given database.
}
\note{
Ninjas are mammals.
}
\examples{
cnn <- get.connection('<server_name>','<database_name>')
doc <- get.documentation(cnn, 'c:/' , 'documentation.pdf')
}
\keyword{tsql}
