% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.markdown.R
\name{get.markdown}
\alias{get.markdown}
\title{Return markdown for database}
\usage{
get.markdown(connection)
}
\arguments{
\item{connection}{ODBC Connection Class}
}
\description{
This function returns a data frame of ordered markdown that can be used to knit a document.
}
\examples{
cnn <- get.connection('<server_name>','<database_name>')
md <- get.markdown(cnn)
View(md)
}
\keyword{tsql}
